🐛 fix: 저장되는 좌표 타입 double로 변경

See merge request s10-final/S10P31A202!85

+++ server/src/main/java/com/zzuiksa/server/domain/route/data/LatLngDto.java
@@ -11,12 +11,12 @@ import lombok.NoArgsConstructor;
 public class LatLngDto {
 
     @NotNull
-    private Float lat;
+    private Double lat;
 
     @NotNull
-    private Float lng;
+    private Double lng;
 
-    public static LatLngDto of(float lat, float lng) {
+    public static LatLngDto of(double lat, double lng) {
         return new LatLngDto(lat, lng);
     }
 }
